Refactored statement and schedule layouts to remove duplicate code: created getSummary for common data, refactored statement and schedule index files to remove the relevant code and call getSummary
