--- app/generator/content/schedule/part1/get-agreement-reference.js
@@ -1,12 +1,0 @@
-const getAgreementReference = (agreementReference) => {
-  return {
-    columns: [
-      { width: 200, text: 'Agreement number:' },
-      { width: '*', text: agreementReference }
-    ],
-    style: 'column',
-    columnGap: 10
-  }
-}
-
-module.exports = getAgreementReference
