--- app/generator/content/schedule/part1/get-sbi.js
@@ -1,12 +1,0 @@
-const getSBI = (sbi) => {
-  return {
-    columns: [
-      { width: 200, text: 'SBI:' },
-      { width: '*', text: sbi }
-    ],
-    style: 'column',
-    columnGap: 10
-  }
-}
-
-module.exports = getSBI
