--- app/generator/content/statement/part1/get-agreement-number.js
@@ -1,12 +1,0 @@
-const getAgreementNumber = (agreementNumber) => {
-	return {
-		columns: [
-			{ width: 200, text: 'Agreement number:' },
-			{ width: '*', text: agreementNumber }
-		],
-		style: 'column',
-		columnGap: 10
-	};
-};
-
-module.exports = getAgreementNumber;
