+++ app/generator/content/statement/summary/index.js
@@ -1,28 +1,10 @@
 const path = require('path')
-const { millimetresToPoints } = require('../../../conversion')
-const getAddress = require('./get-address')
-const getBusinessName = require('../../get-business-name')
-const getSBI = require('./get-sbi')
-const getAgreementNumber = require('../part1/get-agreement-number')
-const imagePath = path.join(__dirname, '../../..', 'images')
+const getSummary = require('../../get-summary')
 
 const summary = (statement) => {
 	return {
 		stack: [
-			{
-				image: `${imagePath}/v2/logo.jpg`,
-				fit: [millimetresToPoints(200), millimetresToPoints(25)],
-				style: 'logo'
-			},
-			getAddress(statement.businessName, statement.address),
-			{
-				text: `${statement.scheme.name} ${statement.scheme.year}`,
-				style: 'header14'
-			},
-			{ text: 'Payment statement', style: 'header14' },
-			getBusinessName(statement.businessName),
-			getSBI(statement.sbi),
-			getAgreementNumber(statement.scheme.agreementNumber),
+			getSummary(summary),
 			`\n\nThis statement explains your payment for the ${statement.scheme.name} (${statement.scheme.shortName}). It is made up of 3 parts.`,
 			'\nPart 1 provides a summary of the most recent payment.',
 			'Part 2 explains how we calculated the payment.',
@@ -32,4 +14,4 @@ const summary = (statement) => {
 	};
 };
 
-module.exports = summary;
+module.exports = statement;
